refactor: move keywords import from doi to software keywords component
